Login Page Protoype
The application now have a login page.
The login page include typing in email and password. Then a when sending, a send function will be called.

A create new account have also been implemented. It is almost the same structure as login page, but with an extra confirm password text field. And This page also triggers a send function.

Needed to be done:
Develop the trigger sending functions when communication to database are up and running.